feat: add consumption history fetching

+++ src/main/java/by/bsuir/caloriestracker/models/ConsumedProduct.java
@@ -20,5 +20,5 @@ public class ConsumedProduct {
     private User user;
     @ManyToOne
     private Product product;
-    private float weight;
+    private int weight;
 }
